requirements.txt
psycopg2-binary
pandas
numpy
SQLAlchemy
python-dotenv
ruff
pytest
